Add syzln.
This is a shell script to generate and maintain the tree of symlinks
that live under the .sf directory, and point back to the actual
syzfiles.

Based on initial work by Dale Roberts <ds.roberts@anu.edu.au> (who also
provided the inspiration for the name), although I'm afraid not much of
the original remains.